thread: suppress thread poll WARNLOGs under gdb
These messages are very helpful to catch cases where a thread gets stuck
for a long time but can be annoying when running the program under gdb.
To suppress them, they're now only printed under a check for a variable
that gets set by a simple .gdbinit script and cleared when detaching
from the process.

Change-Id: Ia98b555443769a3ffbf23dc293f94b4dac8266aa
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>