bdev_qos/burst: Remove duplication between local bucket init and reset
Local bucket does not have capacity now but will have it in the
following patches. Capacity will be fixed at creation. Hence, it will be
necessary to separate local bucket initialization and reset. We already
have such separation but operations are duplicated. Remove the
duplication in this patch as a preparation.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>